Split into smaller files
